transferred getReplies and createReplies to comments controller, and routes
