+++ src/controllers/comment.controller.ts
@@ -2,6 +2,7 @@ import { Request, Response, NextFunction } from "express";
 import catchAsync from "../utils/catchAsync";
 import AppError from "../utils/appError";
 import CommentServices from "../services/comment.service";
+import ReplyServices from "../services/reply.service";
 
 const getComment = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
@@ -71,4 +72,46 @@ const deleteComment = catchAsync(
   },
 );
 
-export { getComment, updateComment, deleteComment };
+const getReplies = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { commentId } = req.params;
+
+    if (!commentId) {
+      return next(new AppError("Comment ID is missing", 400));
+    }
+
+    const comment = await ReplyServices.getRepliesByCommentId(
+      commentId,
+      req.query,
+    );
+
+    if (!comment) {
+      return next(
+        new AppError("Comment does not exist or has been deleted", 404),
+      );
+    }
+
+    res.status(200).json({ status: "Success", data: { comment } });
+  },
+);
+
+const createReply = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { commentId } = req.params;
+    const { text } = req.body;
+
+    const reply = await ReplyServices.createReply(
+      commentId,
+      text,
+      req.user._id,
+    );
+
+    if (!reply) {
+      return next(new AppError("Failed to post reply", 400));
+    }
+
+    res.status(201).json({ status: "Success", data: { reply } });
+  },
+);
+
+export { getComment, updateComment, deleteComment, getReplies, createReply };
